requirements.txt
streamlit==1.46.1
streamlit-authenticator==0.4.2
bcrypt>=4.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
python-docx>=0.8.11
langchain-community>=0.3.0
openai>=1.0.0
fire>=0.5.0
pinecone>=2.0.0
langchain-openai>=0.3.0
langchain-text-splitters>=0.3.0
langchain>=0.3.0
docx2txt>=0.9
langchain-core>=0.3.0
tiktoken>=0.5.0
plotly>=5.0.0
pyyaml>=6.0
supabase>=2.0.0

reflex==0.8.2